Move default config/data path helpers to core/config
desktop/main.go's only commands dependency was DefaultConfigPath; the
XDG probing helpers belong with config loading. Drops internal/commands
and the entire TUI tree from the desktop binary's dependency closure.
