+++ main.go
@@ -182,14 +182,14 @@ Run 'hive new' to create a new session from the current repository.`,
 				Aliases:     []string{"c"},
 				Usage:       "path to config file",
 				Sources:     cli.EnvVars("HIVE_CONFIG"),
-				Value:       commands.DefaultConfigPath(),
+				Value:       config.DefaultConfigPath(),
 				Destination: &flags.ConfigPath,
 			},
 			&cli.StringFlag{
 				Name:        "data-dir",
 				Usage:       "path to data directory",
 				Sources:     cli.EnvVars("HIVE_DATA_DIR"),
-				Value:       commands.DefaultDataDir(),
+				Value:       config.DefaultDataDir(),
 				Destination: &flags.DataDir,
 			},
 		},
